requirements.txt
# PyPhotoManager Dependencies
# Core GUI Framework
PyQt6>=6.5.0

# Image Processing
Pillow>=10.0.0

# Configuration Management
PyYAML>=6.0.0

# Database Utilities
sqlite-utils>=3.35.0

# Development Dependencies
pytest>=7.0.0
pytest-cov>=4.0.0
pytest-qt>=4.2.0
black>=23.0.0
mypy>=1.0.0

# Documentation
sphinx>=5.0.0
sphinx-rtd-theme>=1.2.0

# Packaging
PyInstaller>=5.0.0
setuptools>=65.0.0
wheel>=0.38.0

# Logging and Monitoring
structlog>=22.0.0

# Translation Support
googletrans-py>=4.0.0

# Async Support
asyncio-mqtt>=0.11.0
aiofiles>=22.0.0

# GPS Location Plugin Dependencies
requests>=2.28.0
geopy>=2.3.0

# Testing Utilities
factory-boy>=3.2.0
faker>=18.0.0